Refactor: move `BlogDetailsViewController` extension to its own file
