+++ WordPress/Classes/ViewRelated/Jetpack/Branding/Menu Card/BlogDetailsViewController+JetpackBrandingMenuCard.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+extension BlogDetailsViewController {
+
+    @objc func jetpackCardSectionViewModel() -> BlogDetailsSection {
+        let row = BlogDetailsRow()
+        row.callback = {}
+
+        let section = BlogDetailsSection(title: nil,
+                                         rows: [row],
+                                         footerTitle: nil,
+                                         category: .jetpackBrandingCard)
+        return section
+    }
+}
